AsylumSummaryController - A simple controller for the new summary endpoint of the backend
SummaryQueryParameterDto - The DTO which contains query options for the search

AsylumSummaryModel - The model of the smallest piece of data shows information regarding citizenship, office, amount granted, amount admin closed, amount denied, and totaal cases for that office/citizenship

AsylumYearSummaryModel - The model for a year of information shows overall data same as AsylumSummaryModel, also contains a list of AsylumSummaryModels

AsylumSummaryDto - DTO sent back from API endpoint contains a list of AsylumYearSummaryModel

AsylumSummaryService - The service class handling all the business logic entry point at getSummaryBy. The class first queries the repository with the given query conditions then sends that iterable to the getYears method. The getYears method, (this one is a doozy and may want to be broken down into submethods at some point). First two Hashmaps are created and filled with dummy information, the iterable of information is then iterated across recording information into the hashmaps on whether the case is granted or denied or closed by admin it also store information on what office/citizenship this case correlated to. Finally those Hashmaps are unraveled and placed into their handle little class models and shipped back to the getYears method to be returned to the user.

ToDo
Write test for classes
add functionality to view as percent or raw data